# Conflicts:
#	broker/pom.xml
#	broker/src/main/java/org/apache/rocketmq/broker/BrokerController.java
#	broker/src/main/java/org/apache/rocketmq/broker/latency/BrokerFastFailure.java
#	broker/src/main/java/org/apache/rocketmq/broker/out/BrokerOuterAPI.java
#	broker/src/main/java/org/apache/rocketmq/broker/processor/AdminBrokerProcessor.java
#	broker/src/main/java/org/apache/rocketmq/broker/slave/SlaveSynchronize.java
#	broker/src/main/java/org/apache/rocketmq/broker/topic/TopicConfigManager.java
#	client/pom.xml
#	client/src/main/java/org/apache/rocketmq/client/ClientConfig.java
#	client/src/main/java/org/apache/rocketmq/client/consumer/DefaultMQPullConsumer.java
#	client/src/main/java/org/apache/rocketmq/client/impl/MQClientAPIImpl.java
#	client/src/main/java/org/apache/rocketmq/client/impl/consumer/DefaultMQPullConsumerImpl.java
#	client/src/main/java/org/apache/rocketmq/client/producer/DefaultMQProducer.java
#	common/pom.xml
#	common/src/main/java/org/apache/rocketmq/common/BrokerConfig.java
#	common/src/main/java/org/apache/rocketmq/common/MQVersion.java
#	common/src/main/java/org/apache/rocketmq/common/protocol/RequestCode.java
#	distribution/pom.xml
#	example/pom.xml
#	filter/pom.xml
#	filtersrv/pom.xml
#	logappender/pom.xml
#	namesrv/pom.xml
#	openmessaging/pom.xml
#	pom.xml
#	remoting/pom.xml
#	srvutil/pom.xml
#	store/pom.xml
#	store/src/main/java/org/apache/rocketmq/store/CommitLog.java
#	store/src/main/java/org/apache/rocketmq/store/DefaultMessageStore.java
#	store/src/main/java/org/apache/rocketmq/store/MappedFileQueue.java
#	store/src/test/java/org/apache/rocketmq/store/MappedFileQueueTest.java
#	test/pom.xml
#	tools/pom.xml
#	tools/src/main/java/org/apache/rocketmq/tools/command/consumer/ConsumerProgressSubCommand.java

+++ store/src/main/java/org/apache/rocketmq/store/DefaultMessageStore.java
@@ -42,6 +42,8 @@ import org.apache.rocketmq.common.SystemClock;
 import org.apache.rocketmq.common.ThreadFactoryImpl;
 import org.apache.rocketmq.common.UtilAll;
 import org.apache.rocketmq.common.constant.LoggerName;
+import org.apache.rocketmq.logging.InternalLogger;
+import org.apache.rocketmq.logging.InternalLoggerFactory;
 import org.apache.rocketmq.common.message.MessageDecoder;
 import org.apache.rocketmq.common.message.MessageExt;
 import org.apache.rocketmq.common.message.MessageExtBatch;
@@ -58,14 +60,11 @@ import org.apache.rocketmq.store.schedule.ScheduleMessageService;
 import org.apache.rocketmq.store.stats.BrokerStatsManager;
 import org.apache.rocketmq.store.timer.PerfCounter;
 import org.apache.rocketmq.store.timer.TimerMessageStore;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import static org.apache.rocketmq.store.config.BrokerRole.SLAVE;
 
 public class DefaultMessageStore implements MessageStore {
-
-    private static final Logger log = LoggerFactory.getLogger(LoggerName.STORE_LOGGER_NAME);
+    private static final InternalLogger log = InternalLoggerFactory.getLogger(LoggerName.STORE_LOGGER_NAME);
 
     private final MessageStoreConfig messageStoreConfig;
     // CommitLog
@@ -118,6 +117,8 @@ public class DefaultMessageStore implements MessageStore {
     private final PerfCounter.Ticks ticks =  new PerfCounter.Ticks(log);
 
     private final MessageVersion messageVersion;
+
+    boolean shutDownNormal = false;
 
     public DefaultMessageStore(final MessageStoreConfig messageStoreConfig, final BrokerStatsManager brokerStatsManager,
         final MessageArrivingListener messageArrivingListener, final BrokerConfig brokerConfig) throws IOException {
@@ -275,8 +276,9 @@ public class DefaultMessageStore implements MessageStore {
             this.storeCheckpoint.flush();
             this.storeCheckpoint.shutdown();
 
-            if (this.runningFlags.isWriteable()) {
+            if (this.runningFlags.isWriteable() && dispatchBehindBytes() == 0) {
                 this.deleteFile(StorePathConfigHelper.getAbortFile(this.messageStoreConfig.getStorePathRootDir()));
+                shutDownNormal = true;
             } else {
                 log.warn("the store may be wrong, so shutdown abnormally, and keep abort file.");
             }
@@ -512,7 +514,7 @@ public class DefaultMessageStore implements MessageStore {
                         long maxPhyOffsetPulling = 0;
 
                         int i = 0;
-                        final int maxFilterMessageCount = Math.max(16000, maxMsgNums * ConsumeQueue.CQ_STORE_UNIT_SIZE);
+                        final int maxFilterMessageCount = Math.max(3200, maxMsgNums * ConsumeQueue.CQ_STORE_UNIT_SIZE);
                         final boolean diskFallRecorded = this.messageStoreConfig.isDiskFallRecorded();
                         ConsumeQueueExt.CqExtUnit cqExtUnit = new ConsumeQueueExt.CqExtUnit();
                         for (; i < bufferConsumeQueue.getSize() && i < maxFilterMessageCount; i += ConsumeQueue.CQ_STORE_UNIT_SIZE) {
